+++ .gitignore
@@ -338,3 +338,4 @@ ASALocalRun/
 
 # BeatPulse healthcheck temp database
 healthchecksdb
+/UI/Spotify/polymusic.db
