+++ Common/Spotify.Domain/Entities/Intermediate/AlbumAuthor.cs
@@ -1,19 +1,16 @@
-﻿using System.ComponentModel.DataAnnotations;
-using System.ComponentModel.DataAnnotations.Schema;
-
-namespace Spotify.Domain.Entities
+﻿namespace Spotify.Domain.Entities.Intermediate
 {
-	/// <summary>
-	/// Ассоциативная сущность для <see cref="Spotify.Domain.Entities.Author"/> и <see cref="Spotify.Domain.Entities.Album"/>
-	/// </summary>
-	public class AlbumAuthor
-	{
-		public int AuthorId { get; set; }
+    /// <summary>
+    /// Ассоциативная сущность для <see cref="Entities.Author"/> и <see cref="Entities.Album"/>
+    /// </summary>
+    public class AlbumAuthor
+    {
+        public int AuthorId { get; set; }
 
-		public Author Author { get; set; }
+        public Author Author { get; set; }
 
-		public int AlbumId { get; set; }
+        public int AlbumId { get; set; }
 
-		public Album Album { get; set; }
-	}
+        public Album Album { get; set; }
+    }
 }
