+++ Common/Spotify.Domain/Entities/Intermediate/PlaylistTrack.cs
@@ -1,9 +1,4 @@
-﻿using Spotify.Domain.Entities.Identity;
-using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace Spotify.Domain.Entities.Intermediate
+﻿namespace Spotify.Domain.Entities.Intermediate
 {
     public class PlaylistTrack
     {
