+++ Common/Spotify.Domain/Entities/Intermediate/UserLikedAlbum.cs
@@ -1,18 +1,15 @@
 ﻿using Spotify.Domain.Entities.Identity;
-using System;
-using System.Collections.Generic;
-using System.Text;
 
-namespace Spotify.Domain.Entities
+namespace Spotify.Domain.Entities.Intermediate
 {
-	public class UserLikedAlbum
-	{
-		public int UserId { get; set; }
+    public class UserLikedAlbum
+    {
+        public int UserId { get; set; }
 
-		public User User { get; set; }
+        public User User { get; set; }
 
-		public int AlbumId { get; set; }
+        public int AlbumId { get; set; }
 
-		public Album Album { get; set; }
-	}
+        public Album Album { get; set; }
+    }
 }
