+++ Common/Spotify.Domain/Entities/Intermediate/UserLikedAuthor.cs
@@ -1,20 +1,18 @@
-﻿using Spotify.Domain.Entities.Base;
-using Spotify.Domain.Entities.Identity;
-using System.ComponentModel.DataAnnotations;
+﻿using Spotify.Domain.Entities.Identity;
 
-namespace Spotify.Domain.Entities
+namespace Spotify.Domain.Entities.Intermediate
 {
-	/// <summary>
-	/// Ассоциативная сущность для <see cref="Spotify.Domain.Entities.Author"/> и <see cref="Spotify.Domain.Entities.Identity.User"/>
-	/// </summary>
-	public class UserLikedAuthor
-	{
-		public int UserId { get; set; }
+    /// <summary>
+    /// Ассоциативная сущность для <see cref="Entities.Author"/> и <see cref="Identity.User"/>
+    /// </summary>
+    public class UserLikedAuthor
+    {
+        public int UserId { get; set; }
 
-		public User User { get; set; }
+        public User User { get; set; }
 
-		public int AuthorId { get; set; }
+        public int AuthorId { get; set; }
 
-		public Author Author { get; set; }
-	}
+        public Author Author { get; set; }
+    }
 }
