+++ Common/Spotify.Domain/Entities/Intermediate/UserPlaylist.cs
@@ -1,20 +1,18 @@
-﻿using Spotify.Domain.Entities.Base;
-using Spotify.Domain.Entities.Identity;
-using System.ComponentModel.DataAnnotations;
+﻿using Spotify.Domain.Entities.Identity;
 
-namespace Spotify.Domain.Entities
+namespace Spotify.Domain.Entities.Intermediate
 {
-	/// <summary>
-	/// Ассоциативная сущность для <see cref="Spotify.Domain.Entities.Playlist"/> и <see cref="Spotify.Domain.Entities.Identity.User"/>
-	/// </summary>
-	public class UserPlaylist
-	{
-		public int UserId { get; set; }
+    /// <summary>
+    /// Ассоциативная сущность для <see cref="Entities.Playlist"/> и <see cref="Identity.User"/>
+    /// </summary>
+    public class UserPlaylist
+    {
+        public int UserId { get; set; }
 
-		public User User { get; set; }
+        public User User { get; set; }
 
-		public int PlaylistId { get; set; }
+        public int PlaylistId { get; set; }
 
-		public Playlist Playlist { get; set; }
-	}
+        public Playlist Playlist { get; set; }
+    }
 }
