+++ Common/Spotify.Domain/Entities/Intermediate/UserSubscription.cs
@@ -1,18 +1,15 @@
 ﻿using Spotify.Domain.Entities.Identity;
-using System;
-using System.Collections.Generic;
-using System.Text;
 
-namespace Spotify.Domain.Entities
+namespace Spotify.Domain.Entities.Intermediate
 {
-	public class UserSubscription
-	{
-		public int UserId { get; set; }
+    public class UserSubscription
+    {
+        public int UserId { get; set; }
 
-		public User User { get; set; }
+        public User User { get; set; }
 
-		public int SubscriptionId { get; set; }
+        public int SubscriptionId { get; set; }
 
-		public User Subscription { get; set; }
-	}
+        public User Subscription { get; set; }
+    }
 }
